requirements.txt
mysql-connector-python
sqlalchemy
flask
python-jose[cryptography]
PyJWT
fastapi
uvicorn
passlib[bcrypt]
python-dotenv
pydantic[email]
python-multipart
